move redirection notification modal files to the redirect_notification folder
