+++ src/articles/ArticlePreview.js
@@ -3,7 +3,7 @@ import { useState } from "react";
 import moment from "moment";
 import { isMobile } from "../utils/misc/browserDetection";
 import * as s from "./ArticlePreview.sc";
-import RedirectionNotificationModal from "../components/RedirectionNotificationModal";
+import RedirectionNotificationModal from "../components/redirect_notification/RedirectionNotificationModal";
 import Feature from "../features/Feature";
 import { extractVideoIDFromURL } from "../utils/misc/youtube";
 import SmallSaveArticleButton from "./SmallSaveArticleButton";
