+++ src/components/redirect_notification/RedirectionNotificationForSupportedBrowsers.js
